+++ src/(domain)/book/summary-widget/hooks/useWidgetItem.ts
