+++ src/(domain)/book/summary-widget/util/widget.ts
